objektdb_core::storage_engine::file_manager

Function create_db

pub fn create_db(db_name: String) -> Result<(), String>

Creates a new database from the name.

It initially creates it empty, and you can then insert tables using the create_table() method. The binary file that is created by the method follows the following format:

DATABASE_HEADER {
   magic_number = 0x4D594442, (4 byte(u32))
   version = 1, (1 byte(u8))
   num_tablespaces = N, (1 byte(u8))
   flags = u32   (4 byte)
}

TABLESPACE_DIRECTORY[N] {
   TablespaceEntry {
       name = "User", (64 byte)
       file_path = "user.tbl", (68 byte)
       offset = 123, (4 byte(u32))
       checksum = u32, (4 byte(u32))
       last_oid = u64,  (8 byte(u64))
   }
   ...
}

Arguments

db_name - The name of the database (without the .db extension).

Returns

Ok(()) if the database file was successfully created.

Err(String) if the file does not exist or if an error occurred during creation.

Example

use objektoDB::storage_engine::file_manager::create_db;

match create_db(String::from("my_database")) {
    Ok(_) => println!("Database created successfully!"),
    Err(e) => println!("Error creating database: {}", e),
}
If the conversion is successful, it returns `Ok(())`, otherwise `Err(error)`